Fix tests and clean up function locations
